Entities/src/main/java/be/groups/glanguage/glanguage/api/entities/formula/implementations/binary/FormulaIntegerDivision.java
package be.groups.glanguage.glanguage.api.entities.formula.implementations.binary;

import be.groups.glanguage.glanguage.api.entities.evaluation.Evaluator;
import be.groups.glanguage.glanguage.api.entities.formula.AbstractFormula;
import be.groups.glanguage.glanguage.api.entities.formula.description.FormulaDescription;
import be.groups.glanguage.glanguage.api.entities.formula.description.FormulaType;
import com.fasterxml.jackson.annotation.JsonIgnore;

import javax.persistence.DiscriminatorValue;
import javax.persistence.Entity;
import javax.persistence.Transient;

@Entity
@DiscriminatorValue(FormulaType.Values.OP_INTEGER_DIVISION)
public class FormulaIntegerDivision extends BinaryFormula {

	protected FormulaIntegerDivision() {
		super();
	}

	public FormulaIntegerDivision(FormulaDescription description, AbstractFormula child1, AbstractFormula child2) {
		super(description, child1, child2);
	}

	@JsonIgnore
	@Transient
	@Override
	public Integer getIntegerValue(Evaluator evaluator) {
		return getParameters().get(0).getIntegerValue(evaluator) / getParameters().get(1).getIntegerValue(evaluator);
	}

	@Override
	public String operationAsText() {
		return "/";
	}

}
